refactor: move feide db functions to server folder
