+++ src/auth/authoptions.ts
@@ -1,7 +1,6 @@
 import FeideProvider from './feide/FeideProvider'
 import PrismaAdapter from './feide/PrismaAdapter'
 import signUp from './feide/signUp'
-import { updateFeideAccount } from './feide/index'
 import prisma from '@/prisma'
 import { readPermissionsOfUser } from '@/server/rolePermissions/read'
 import { readMembershipsOfUser } from '@/server/groups/read'
@@ -10,6 +9,7 @@ import { decode } from 'next-auth/jwt'
 import type { JWT } from 'next-auth/jwt'
 import type { AuthOptions, Profile, User as nextAuthUser } from 'next-auth'
 import type { ExtendedFeideUser } from './feide/Types'
+import { updateFeideAccount } from '@/server/auth/feide/update'
 
 export const authOptions: AuthOptions = {
     providers: [
